refactor: eliminate CommandDispatcher, library subscribes directly (#521)
Replace the CommandDispatcher middleman with a single
subscribe_commands() function in src/library/commands.rs. The library
now subscribes directly to *Requested events and handles them on Tokio.

- Consolidate 10 command handler files + dispatcher + trait into one
  ~200-line module with a single match block
- Delete entire src/commands/ directory (13 files)
- MockLibrary and all tests ported to the new module
- Doc comments updated to remove CommandDispatcher references
