+++ src/application.rs
@@ -51,8 +51,6 @@ mod imp {
         /// Centralised event bus for fan-out event delivery.
         /// Created when the library is loaded.
         pub event_bus: RefCell<Option<EventBus>>,
-        /// Holds the command dispatcher so its subscription stays alive.
-        pub dispatcher: RefCell<Option<crate::commands::dispatcher::CommandDispatcher>>,
         /// App-lifetime event bus subscriptions (error toasts, etc.).
         pub subscriptions: RefCell<Vec<crate::event_bus::Subscription>>,
     }
@@ -505,19 +503,15 @@ impl MomentsApplication {
                             .clone();
                         window.setup(library, tokio.clone(), settings, &bus);
 
-                        // Create the command dispatcher — routes *Requested
+                        // Subscribe for command events — routes *Requested
                         // events to library calls on the Tokio runtime.
-                        // Stored in the app imp so the subscription stays alive.
                         let Some(lib) = app.imp().library.borrow().as_ref().map(Arc::clone) else {
-                            tracing::error!("library not initialised when creating dispatcher");
+                            tracing::error!("library not initialised when subscribing commands");
                             return;
                         };
-                        let dispatcher = crate::commands::dispatcher::CommandDispatcher::new(
-                            lib,
-                            tokio.clone(),
-                            &bus,
-                        );
-                        *app.imp().dispatcher.borrow_mut() = Some(dispatcher);
+                        let cmd_sub =
+                            crate::library::commands::subscribe_commands(lib, tokio.clone(), &bus);
+                        app.imp().subscriptions.borrow_mut().push(cmd_sub);
 
                         // Subscribe for error toasts — centralised error
                         // handling for all command failures.
